% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfs.R
\name{mf.perm.resid}
\alias{mf.perm.resid}
\title{The residual from linear model (mf.resid) and the permuted one (mf.perm.resid).}
\usage{
mf.perm.resid(data = data_long.vs.Jnt, n.perm = 500,
  model = my.model, var = "Visit")
}
\arguments{
\item{data}{A dataframe with column of (to be imput to mf.resid function as data_long.vs.Jnt argument)}

\item{n.perm}{The number of permuted data.}

\item{model}{A formula for lm() function to calculate residuals.}

\item{var}{A character vector to specify the categorical variate to differring the residuals.}
}
\description{
The residual from linear model (mf.resid) and the permuted one (mf.perm.resid).
}
