% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfs.R
\name{mf.resid}
\alias{mf.resid}
\title{The residual from linear model (mf.resid)}
\usage{
mf.resid(data_long.vs.Jnt, model = my.model, var = "Visit")
}
\arguments{
\item{data_long.vs.Jnt}{A dataframe with column of}

\item{model}{A formula for lm() function to calculate residuals.}

\item{var}{A character vector to specify the categorical variate to differring the residuals.}
}
\description{
The residual from linear model (mf.resid)
}
